feat/#108: content 뷰 관련 api 연결
